Move enqueue metadata to job derive
