+++ oxanus/examples/cron_w_err.rs
@@ -11,7 +11,7 @@ enum WorkerError {
 #[derive(Debug, Serialize, Deserialize, Clone)]
 struct WorkerContext {}
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, oxanus::Job)]
 struct TestJob {}
 
 #[derive(oxanus::Worker)]
