+++ oxanus/examples/minimal.rs
@@ -10,7 +10,7 @@ enum WorkerError {}
 #[derive(Debug, Serialize, Deserialize, Clone)]
 struct WorkerContext {}
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, oxanus::Job)]
 struct TestJob {
     sleep_s: u64,
 }
