+++ oxanus/examples/unique.rs
@@ -10,19 +10,19 @@ enum WorkerError {}
 #[derive(Debug, Serialize, Deserialize, Clone)]
 struct WorkerContext {}
 
-#[derive(Debug, Serialize, Deserialize)]
-struct Worker2SecJob {
+#[derive(Debug, Serialize, Deserialize, oxanus::Job)]
+#[oxanus(unique_id = "worker2sec:{id}", on_conflict = Skip)]
+struct TwoSecJob {
     id: usize,
 }
 
 #[derive(oxanus::Worker)]
-#[oxanus(unique_id = "worker2sec:{id}", on_conflict = Skip)]
-struct Worker2Sec;
+struct TwoSecWorker;
 
-impl Worker2Sec {
+impl TwoSecWorker {
     async fn process(
         &self,
-        _job: &Worker2SecJob,
+        _job: &TwoSecJob,
         _ctx: &oxanus::JobContext,
     ) -> Result<(), WorkerError> {
         tokio::time::sleep(std::time::Duration::from_millis(2000)).await;
@@ -45,9 +45,9 @@ pub async fn main() -> Result<(), oxanus::OxanusError> {
     let storage = oxanus::Storage::builder().build_from_env()?;
     let config = ComponentRegistry::build_config(&storage);
 
-    storage.enqueue(QueueOne, Worker2SecJob { id: 1 }).await?;
-    storage.enqueue(QueueOne, Worker2SecJob { id: 1 }).await?;
-    storage.enqueue(QueueOne, Worker2SecJob { id: 2 }).await?;
+    storage.enqueue(QueueOne, TwoSecJob { id: 1 }).await?;
+    storage.enqueue(QueueOne, TwoSecJob { id: 1 }).await?;
+    storage.enqueue(QueueOne, TwoSecJob { id: 2 }).await?;
 
     oxanus::run(config, ctx).await?;
